minor improvements

+++ app/src/main/java/com/service_kluch/nmedia/dto/Post.kt
@@ -6,7 +6,7 @@ data class Post(
     val content: String,
     val published: String,
     val likeByMe: Boolean = false,
-    val likeCount: Int = 0,
-    val shareCount: Int = 0,
-    val watchesCount: Int = 0
+    val likeCount: Long = 0,
+    val shareCount: Long = 0,
+    val watchesCount: Long = 0
 )
